Use Dropzone item cache to store drop target.
This is an imperfect but adequate way to deal with uploading to multiple
folders at the same time. It's safe to cache the drop target in the drop
event and apply it in the added file event as long as users don't drop
onto different folders at superhuman speeds, or drop very large numbers
of files at once. Still, we should find a better fix for this at some
point.